[Refactor] loadView() 통해 View와 ViewController 분리 #64
